converted ArcGisElcUI to TypeScript
