+++ packages/demo/index.html
@@ -2,7 +2,7 @@
 <html xmlns="http://www.w3.org/1999/xhtml">
 <head>
     <title></title>
-    <link href="//js.arcgis.com/3.15/esri/css/esri.css" rel="stylesheet" />
+    <link href="//js.arcgis.com/3.21/esri/css/esri.css" rel="stylesheet" />
     <link href="../../bower_components/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet" />
     <link href="../../bower_components/bootstrap/dist/css/bootstrap-theme.min.css" rel="stylesheet" />
     <link href="../../elc-ui/elc-ui-behavior.css" rel="stylesheet" />
@@ -47,7 +47,7 @@
             window.dojoConfig = dojoConfig;
         }(location.href.replace(/\?.+$/i, "").replace(/\/[^\/]*$/, "")));
     </script>
-    <script src="//js.arcgis.com/3.15/init.js" integrity="sha384-q6MfUBWGKtML65ieXFuzf+nX2/ylnvnDgiFPgIMFfpYduqJ4T6twwq3vqhMu29Mt" crossorigin="anonymous"></script>
+    <script src="//js.arcgis.com/3.21/init.js" integrity="sha384-y9l0OAZ2wUBSau+YStQwIsOkyJTpTyzuqulXd/g0xThdE8RfIkrVhNIyKRVxgv2+" crossorigin="anonymous"></script>
     <script src="Script/index.js"></script>
 </body>
 </html>
