+++ packages/demo/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "demo",
+  "version": "1.0.0",
+  "main": "index.js",
+  "private": true,
+  "license": "Unlicense",
+  "dependencies": {
+    "ts-loader": "^4.3.0",
+    "tslib": "^1.9.1",
+    "webpack": "^4.10.1",
+    "webpack-cli": "^2.1.4"
+  }
+}
