--- packages/elc-ui/Templates/elc-ui-bootstrap.min.html
@@ -1,1 +1,0 @@
-<!DOCTYPE html><html><head><title>ELC UI Template</title></head><body><div class="elc-ui-root"><div role="tabpanel"><ul class="nav nav-tabs" role="tablist"><li role="presentation" class="active"><a href="#findRouteLocationSection" role="tab" data-toggle="tab" title="Find Route Location"><span class="glyphicon glyphicon-edit"></span><span class="sr-only">Find Route Location</span></a></li><li role="presentation"><a href="#findNearestRouteLocationSection" role="tab" data-toggle="tab" title="Nearest Route Location"> <span class="glyphicon glyphicon-screenshot"></span> <span class="sr-only">Nearest Route Location</span> </a></li><li role="presentation"><a href="#optionsSection" role="tab" data-toggle="tab" title="Options"> <span class="glyphicon glyphicon-cog"></span> <span class="sr-only">Options</span> </a></li></ul><div class="tab-content"><section class="find-route-location-section tab-pane active" id="findRouteLocationSection" role="tabpanel"><form action="." class="find-route-location-form mp-mode-srmp geo-mode-point"><div class="form-group"><label>Route</label><div class="form-group"><div class="input-group route-selector"></div></div></div><div class="radio-group form-group"><label>Milepost Type</label> <label><input type="radio" name="milepostType" value="SRMP" checked> <abbr title="State Route Milepost">SRMP</abbr></label> <label><input type="radio" name="milepostType" value="ARM"> <abbr title="Accumulated Route Mileage">ARM</abbr></label></div><div class="radio-group form-group"><label>Geometry Type</label> <label><input type="radio" name="geometryType" value="point" checked> Point</label> <label><input type="radio" name="geometryType" value="line"> Line</label></div><div class="form-group"><label>Reference Date</label> <input name="referenceDate" type="date" required class="form-control"></div><div class="form-group"><div class="start-milepost-section"><label><span class="for-line-only">Start </span>Milepost</label><div class="input-group"><input name="milepost" type="number" required min="0" step="0.01" class="form-control"> <label class="for-srmp-only input-group-addon"><input type="checkbox" name="back"> Back</label></div></div><div class="end-milepost-section for-line-only"><label>End Milepost</label><div class="input-group"><input name="endMilepost" type="number" min="0" step="0.01" class="form-control"> <label class="for-srmp-only input-group-addon"><input type="checkbox" name="endBack"> Back</label></div></div></div><div class="btn-group btn-group-justified" role="group"><div class="btn-group" role="group"><button type="submit" class="btn btn-primary">Find</button></div><div class="btn-group" role="group"><button type="reset" class="btn btn-default">Reset</button></div></div></form></section><section class="find-nearest-route-location-section tab-pane" id="findNearestRouteLocationSection" role="tabpanel"><form action="." method="post" class="find-nearest-route-location-form"><label>Search Radius (in feet)</label> <input name="radius" type="number" required value="200" min="0" title="Determines how far around the clicked point to search for a route." class="form-control"><div class="btn-group btn-group-justified"><div class="btn-group"><button type="submit" class="btn btn-primary">Find</button></div></div></form></section><section class="options-section tab-pane" role="tabpanel" id="optionsSection"><h1>Options</h1><div class="btn-group btn-group-justified"><div class="btn-group"><button type="button" class="clear-graphics-button btn btn-warning" title="Clears graphics created by this tool.">Clear Graphics</button></div></div></section></div></div></div></body></html>
