--- packages/elc-ui/Templates/elc-ui.min.html
@@ -1,1 +1,0 @@
-<!DOCTYPE html><html><head><title>ELC UI Template</title></head><body><div class="elc-ui-root"><section class="find-route-location-section"><h1>Find Route Location</h1><form action="." class="find-route-location-form mp-mode-srmp geo-mode-point"><div><label>Route</label><div class="flex-row route-selector"></div></div><div class="radio-group"><label>Milepost Type</label> <label><input type="radio" name="milepostType" value="SRMP" checked> <abbr title="State Route Milepost">SRMP</abbr></label> <label><input type="radio" name="milepostType" value="ARM"> <abbr title="Accumulated Route Mileage">ARM</abbr></label></div><div class="radio-group"><label>Geometry Type</label> <label><input type="radio" name="geometryType" value="point" checked> Point</label> <label><input type="radio" name="geometryType" value="line"> Line</label></div><div><label>Reference Date</label> <input name="referenceDate" type="date" required></div><div><div class="start-milepost-section"><label><span class="for-line-only">Start </span>Milepost</label><div class="flex-row"><input name="milepost" type="number" required min="0" step="0.01"> <label class="for-srmp-only"><input type="checkbox" name="back"> Back</label></div></div><div class="end-milepost-section for-line-only"><label>End Milepost</label><div class="flex-row"><input name="endMilepost" type="number" min="0" step="0.01"> <label class="for-srmp-only"><input type="checkbox" name="endBack"> Back</label></div></div></div><div class="button-group"><button type="submit">Find</button> <button type="reset">Reset</button></div></form></section><section class="find-nearest-route-location-section"><h1>Find Nearest Route Location</h1><form action="." method="post" class="find-nearest-route-location-form"><div><label>Search Radius (in feet)</label> <input name="radius" type="number" required value="200" min="0" title="Determines how far around the clicked point to search for a route."></div><button type="submit">Find</button></form></section><section class="options-section"><h1>Options</h1><button type="button" class="clear-graphics-button" title="Clears graphics created by this tool.">Clear Graphics</button></section></div></body></html>
