--- packages/elc-ui/demo/ArcGisElcUI/Script/custom-event-polyfill.js
@@ -1,23 +1,0 @@
-// Polyfill for creating CustomEvents on IE9/10/11
-
-// code pulled from:
-// https://github.com/d4tocchini/customevent-polyfill
-// https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent#Polyfill
-
-if (!window.CustomEvent || typeof window.CustomEvent !== 'function') {
-	var CustomEvent = function (event, params) {
-		var evt;
-		params = params || {
-			bubbles: false,
-			cancelable: false,
-			detail: undefined
-		};
-
-		evt = document.createEvent("CustomEvent");
-		evt.initCustomEvent(event, params.bubbles, params.cancelable, params.detail);
-		return evt;
-	};
-
-	CustomEvent.prototype = window.Event.prototype;
-	window.CustomEvent = CustomEvent; // expose definition to window
-}
