--- packages/elc-ui/index.d.ts
@@ -1,10 +1,0 @@
-declare module "dojo/text!./Templates/elc-ui-bootstrap.min.html" {
-    const text: string;
-    export = text;
-}
-
-declare module "dojo/text!./Templates/elc-ui.min.html" {
-    const text: string;
-    export = text;
-}
-
