--- packages/elc-ui/tslint.json
@@ -1,24 +1,0 @@
-{
-    "defaultSeverity": "warn",
-    "extends": [
-        "tslint:recommended"
-    ],
-    "jsRules": {},
-    "rules": {
-        "max-line-length": {
-            "severity": "warn"
-        },
-        "object-literal-sort-keys": false,
-
-        "prefer-const": {
-            "severity": "warn"
-        },
-        "trailing-comma": {
-            "options": {
-                "multiline": "never",
-                "singleline": "never"
-            }
-        }
-    },
-    "rulesDirectory": []
-}
