+++ packages/wsdot-elc/package.json
@@ -2,28 +2,13 @@
   "name": "wsdot-elc",
   "version": "3.1.0",
   "description": "JavaScript client for WSDOT ELC map service SOE",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
-    "index.d.ts",
-    "FormatError.d.ts",
-    "FormatError.js",
-    "lrsTypeUtils.d.ts",
-    "lrsTypeUtils.js",
-    "Route.d.ts",
-    "Route.js",
-    "RouteId.d.ts",
-    "RouteId.js",
-    "RouteLocation.d.ts",
-    "RouteLocation.js",
-    "RouteLocationInterfaces.d.ts",
-    "RouteLocator.d.ts",
-    "routeTypeUtils.d.ts",
-    "routeTypeUtils.js",
-    "routeUtils.d.ts",
-    "routeUtils.js"
+    "dist/*"
   ],
   "directories": {
+    "lib": "dist",
     "test": "spec"
   },
   "scripts": {
@@ -61,11 +46,17 @@
     "karma-jasmine": "^1.1.2",
     "karma-phantomjs-launcher": "^1.0.4",
     "karma-requirejs": "^1.1.0",
+    "prettier": "^1.13.2",
     "requirejs": "^2.3.5",
     "tslint": "^5.10.0",
-    "typescript": "^2.8.3",
-    "prettier": "^1.13.2",
     "tslint-config-prettier": "^1.13.0",
-    "tslint-plugin-prettier": "^1.3.0"
+    "tslint-plugin-prettier": "^1.3.0",
+    "typescript": "^2.8.3"
+  },
+  "dependencies": {
+    "ts-loader": "^4.3.0",
+    "tslib": "^1.9.1",
+    "webpack": "^4.10.1",
+    "webpack-cli": "^2.1.4"
   }
 }
